Fixed GeometryTest failures on macOS
* With hidden visibility by default, dynamic_cast failed for
classes defined only in the header

+++ test/GeometryTest.cpp
@@ -32,7 +32,7 @@ const AbstractGeometry::SharedVertexList* createVertexList(AbstractGeometry* geo
 const AbstractGeometry::SharedEdgeList* createEdgeList(AbstractGeometry* geom)
 {
   auto ds = geom->getDataStructure();
-  auto dataStore = new DataStore<size_t>(2, 0);
+  auto dataStore = new DataStore<uint64_t>(2, 0);
   auto dataArr = ds->createDataArray("Edges", dataStore, geom->getId());
   return dynamic_cast<const AbstractGeometry::SharedEdgeList*>(dataArr);
 }
@@ -40,7 +40,7 @@ const AbstractGeometry::SharedEdgeList* createEdgeList(AbstractGeometry* geom)
 const AbstractGeometry::SharedFaceList* createFaceList(AbstractGeometry* geom)
 {
   auto ds = geom->getDataStructure();
-  auto dataStore = new DataStore<size_t>(4, 0);
+  auto dataStore = new DataStore<uint64_t>(4, 0);
   auto dataArr = ds->createDataArray("Faces", dataStore, geom->getId());
   return dynamic_cast<const AbstractGeometry::SharedFaceList*>(dataArr);
 }
